adds async wiki wrapper
