+++ prompting/tools/datasets/__init__.py
@@ -1,6 +1,7 @@
-from .context import Context
+from .context import Context, BatchContext
 from .base import Dataset
 from .code import HFCodingDataset, StackOverflowDataset
 from .math import MathDataset
 from .mock import MockDataset
 from .wiki import WikiDataset, WikiDateDataset
+from .batch_wiki import BatchWikiDataset
