+++ prompting/utils/__init__.py
@@ -2,5 +2,6 @@ from . import config
 from . import misc
 from . import uids
 from . import logging
-from . import async_wiki
+from . import async_wiki_utils
+from . import async_wiki_wrapper
 from . import wiki
